word-to-md-python/requirements.txt
mammoth
html2text
markdown-it-py
mdit-py-plugins
